renames to viral inj, test resources
