+++ tests/slims/viral_injection/__init__.py
@@ -1,0 +1,1 @@
+"""Package to test Viral Injection data from SLIMS"""
